pyproject.toml
[tool.poetry]
name = "rp-embassy"
version = "0.1.0"
description = ""
authors = ["András Somi <somiandras@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-rtmidi = "^1.5.8"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
